Rename appimage_env to app_image_env
